+++ orbitSimulator/orbitSimulator/position.cpp
@@ -11,7 +11,7 @@
 #include <cassert>
 
 
-Position::Position(float x, float y) : x(0.0), y(0.0)
+Position::Position(double x, double y) : x(0.0), y(0.0)
 {
    setMetersX(x);
    setMetersY(y);
