+++ orbitSimulator/orbitSimulator/spaceObject.h
@@ -22,7 +22,7 @@ class SpaceObject
 public:
 	SpaceObject(): pos(0.0, 0.0), radius(0.0) {}
 	
-	SpaceObject(const Position &pos, float radius):
+	SpaceObject(const Position &pos, double radius):
 	pos(pos), radius(0.0)
 	{
 		// Earth's radius = 6378000 meters
@@ -32,24 +32,24 @@ public:
 			// fragments have smallest radius of 2 px
 			// GPS Satellite has largest radius of 12 px
 		//	assert(radius >= 2 && radius <= 12);
-			float radiusMeters = pos.convertToMeters(radius);
+			double radiusMeters = pos.convertToMeters(radius);
 			setRadius(radiusMeters);
 		}
 	}
 	
-	SpaceObject(float x, float y): pos(x, y), radius(0.0) {}
-	SpaceObject(float x, float y, float radius): pos(x, y), radius(0.0)
+	SpaceObject(double x, double y): pos(x, y), radius(0.0) {}
+	SpaceObject(double x, double y, double radius): pos(x, y), radius(0.0)
 	{
 		if (radius != 6378000)
 		{
 			//assert(radius >= 2 && radius <= 12);
-			float radiusMeters = pos.convertToMeters(radius);
+			double radiusMeters = pos.convertToMeters(radius);
 			setRadius(radiusMeters);
 		}
 	}
 
-	void setRadius(float radius) { this->radius = radius; }
-	void setPosition(float x, float y)
+	void setRadius(double radius) { this->radius = radius; }
+	void setPosition(double x, double y)
 	{
 		pos.setMetersX(x);
 		pos.setMetersY(y);
@@ -58,16 +58,16 @@ public:
 //	virtual void draw() = 0;
 //	virtual void move(int time);
 	
-	float getRadius()      const { return radius; }
+	double getRadius()      const { return radius; }
 	Position getPosition() const { return pos;    }
 
 	Position getPos() { return pos; }
-	float getPosX()  const { return pos.getMetersX(); }
-	float getPosY()  const { return pos.getMetersY(); }
+	double getPosX()  const { return pos.getMetersX(); }
+	double getPosY()  const { return pos.getMetersY(); }
 	
 protected:
 	Position pos;
-	float radius;
+	double radius;
 
 };
